requirements.txt
django-registration






sqlite3
